chore: refactor geolocation
