+++ components/FormikFields/PaperInputPicker/index.js
@@ -1,4 +1,3 @@
-import { Spinner } from 'native-base';
 import * as React from 'react';
 import {
   Image, Text, TouchableWithoutFeedback, View
@@ -8,13 +7,12 @@ import {
   TextInput,
 } from 'react-native-paper';
 
-import getLocation from '../../../modules/geolocation';
 import I18n from '../../../modules/i18n';
 import { layout, theme } from '../../../modules/theme';
-import PaperButton from '../../Button';
 import UseCameraRoll from '../../Multimedia/CameraRoll';
 import UseCamera from '../../Multimedia/UseCamera';
 import AutoFill from './AutoFill';
+import Geolocation from './Geolocation';
 import HouseholdManager from './HouseholdManager';
 import {
   styleButton, styles, stylesDefault, stylesPaper, styleX
@@ -33,21 +31,6 @@ const PaperInputPicker = ({
   const {
     handleChange, handleBlur, errors, setFieldValue, values
   } = formikProps;
-
-  const [location, setLocation] = React.useState({ latitude: 0, longitude: 0, altitude: 0 });
-  const [locationLoading, setLocationLoading] = React.useState(false);
-
-  const handleLocation = async () => {
-    setLocationLoading(true);
-    const currentLocation = await getLocation().catch((e) => e);
-    const { latitude, longitude, altitude } = currentLocation.coords;
-
-    setFieldValue('location', { latitude, longitude, altitude });
-    setLocation({ latitude, longitude, altitude });
-    setTimeout(() => {
-      setLocationLoading(false);
-    }, 1000);
-  };
 
   const translatedLabel = customForm ? label : I18n.t(label);
   const translatedLabelSide = customForm ? sideLabel : I18n.t(sideLabel);
@@ -347,39 +330,11 @@ const PaperInputPicker = ({
         </View>
       )}
       {fieldType === 'geolocation' && (
-        <View key={formikKey}>
-          {location === null && (
-            <PaperButton
-              onPressEvent={handleLocation}
-              buttonText={I18n.t('paperButton.getLocation')}
-            />
-          )}
-          {location !== null && (
-            <View>
-              <PaperButton
-                onPressEvent={handleLocation}
-                buttonText={I18n.t('paperButton.getLocationAgain')}
-              />
-              <View style={{ marginLeft: 'auto', marginRight: 'auto', flexDirection: 'row' }}>
-                {
-                  locationLoading === true
-                  && <Spinner color={theme.colors.primary} />
-                }
-                {locationLoading === false
-                  && (
-                    <View>
-                      <Headline>
-                        {`(${location.latitude.toFixed(2)}, ${location.longitude.toFixed(2)})`}
-                      </Headline>
-                    </View>
-                  )}
-              </View>
-              <Text style={{ color: 'red' }}>
-                {errors[formikKey]}
-              </Text>
-            </View>
-          )}
-        </View>
+        <Geolocation
+          errors={errors}
+          formikKey={formikKey}
+          setFieldValue={setFieldValue}
+        />
       )}
       {fieldType === 'household' && (
         <View key={formikKey}>
